* Added implicit bool conversion to status and status_return * Added interop testing of status and status_return * Added another error value to status_code

+++ unit_tests/test_readers_writer_lock.cpp
@@ -48,5 +48,5 @@ TEST( readers_writer_lock , basic_test )
 		}
 
 	// make sure there were 100 values written (10 threads * 10 writes per thread)
-	ASSERT_EQ( ProtectedValue , 100 );
+	ASSERT_EQ( ProtectedValue , (uint32_t)100 );
 	}
